Refactor, added BaseResizableComponent, BaseNonResizableComponent

+++ GUIFramework/src/BaseComponents/BaseButton.cpp
@@ -7,7 +7,7 @@ using namespace std;
 
 namespace gui_framework
 {
-	BaseButton::BaseButton(const wstring& buttonName, const wstring& buttonText, const utility::ComponentSettings& settings, BaseComponent* parent, uint32_t buttonId, const function<LRESULT(WPARAM, LPARAM)>& onClick) :
+	BaseButton::BaseButton(const wstring& buttonName, const wstring& buttonText, const utility::ComponentSettings& settings, BaseComponent* parent, uint64_t buttonId, const function<LRESULT(WPARAM, LPARAM)>& onClick) :
 		BaseComponent
 		(
 			wstring(standard_classes::button),
@@ -32,7 +32,7 @@ namespace gui_framework
 		return onClick;
 	}
 
-	uint32_t BaseButton::getButtonId() const
+	uint64_t BaseButton::getButtonId() const
 	{
 		return buttonId;
 	}
